Day 05 (initial solution)

+++ AdventOfCode2023.CSharp/ConsoleApp1/Solution.cs
@@ -15,17 +15,13 @@ namespace Aoc2023.ActiveDay
         private List<string> SplitToLines(string input) => Regex.Split(input, NewLine).Where(ln => ln != "").ToList();
         private List<string> Tokenize(string line, IEnumerable<char> splitChars) => line.Split(splitChars.ToArray(), StringSplitOptions.RemoveEmptyEntries).ToList();
 
-        public int Solve1(string input)
+        public long Solve1(string input)
         {
-            //var lns = SplitToLines(input);
-
             return 0;
         }
 
-        public int Solve2(string input)
+        public long Solve2(string input)
         {
-            //var lns = SplitToLines(input);
-
             return 0;
         }
     }
